refactor: move simulat/core/windows to simulat/core/ui/windows
BREAKING CHANGE: move core folder to subfolder
